refactor: import `library.json` instead of `library.ts`
